api/requirements.txt
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
dotenv-python==0.0.1
Flask==2.0.1
Flask-Cors==3.0.10
google-auth==1.30.1
google-auth-oauthlib==0.4.4
gspread==3.7.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.17.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.5
Werkzeug==2.0.1
